qmake: added deployment files support for winrt and winphone targets
The user needs to specify DEPLOYMENT variable. The syntax
is the same as previously used DEPLOYMENT. For more info
please refer to the qmake documentation. The change adds
a new itemgroup, "Deployment Files". All files in this
itemgroup are marked as DeploymentContent and are then
packaged with the application either as XAP or WinRT
specific file format.

Change-Id: I4e92020c068ae31c6ed4fd8f60f0cea122ab3fd2
Reviewed-by: Joerg Bornemann <joerg.bornemann@digia.com>