WinRT: Enable command line passing from main
WinRT passes the executable and Appx server info to the CRT main, and
supports several additional activation arguments as well. This handles the
arguments passed to main as well as the case where a modern app is
launched from an external application (e.g. Qt Creator).
It also simplifies the calculation of applicationFilePath(), which can
now be gleaned from argv[0], which reverts 6af207. Fixes the lack of
command-line passing in the original QPA commit 0647ae.

Finally, the current (.) directory is added to applicationLibPaths for
WinRT because the application directory is not an effective library search
path while the releative path is.

Task-number: QTBUG-30198
Change-Id: Ib4f9eddba9ac5f05d8681d756f3fc9dd41aa57b1
Reviewed-by: Oliver Wolff <oliver.wolff@digia.com>
Reviewed-by: Joerg Bornemann <joerg.bornemann@digia.com>